Fix Edizon, and create upgrade package.

+++ modules.sh
@@ -26,7 +26,7 @@
 #   The version number.
 download_atmosphere () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "Atmosphere-NX" "Atmosphere")
+    latest_release=$(./common.sh get_latest_release "${2}" "Atmosphere-NX" "Atmosphere" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "atmosphere*" "*.zip")
     file=$(./common.sh download_file "${asset}")
@@ -55,7 +55,7 @@ download_atmosphere () {
 #   The version number.
 download_hekate () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${3}" "CTCaer" "hekate")
+    latest_release=$(./common.sh get_latest_release "${3}" "CTCaer" "hekate" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "hekate*" "*.zip")
     file=$(./common.sh download_file "${asset}")
@@ -80,7 +80,7 @@ download_hekate () {
 
 download_appstore () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "vgmoose" "hb-appstore")
+    latest_release=$(./common.sh get_latest_release "${2}" "vgmoose" "hb-appstore" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -93,20 +93,29 @@ download_appstore () {
 
 download_edizon () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "WerWolv" "EdiZon")
+    latest_release=$(./common.sh get_latest_release "${2}" "WerWolv" "EdiZon" "0")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.zip")
-    file=$(./common.sh download_file "${asset}")
+    if [ ! -z "${asset}" ]
+    then
+        file=$(./common.sh download_file "${asset}")
 
-    unzip -qq "${file}" -d "${1}"
-    rm -f "${file}"
+        unzip -qq "${file}" -d "${1}"
+        rm -f "${file}"
+    else
+        asset=$(./common.sh find_asset "${latest_release}" "*.nro")
+        file=$(./common.sh download_file "${asset}")
+
+        mkdir -p "${1}/switch/EdiZon"
+        mv "${file}" "${1}/switch/EdiZon/EdiZon.nro"
+    fi
 
     echo $(./common.sh get_version_number "${latest_release}")
 }
 
 download_emuiibo () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "XorTroll" "emuiibo")
+    latest_release=$(./common.sh get_latest_release "${2}" "XorTroll" "emuiibo" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "emuiibo*" "*.zip")
     file=$(./common.sh download_file "${asset}")
@@ -123,7 +132,7 @@ download_emuiibo () {
 
 download_goldleaf () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "XorTroll" "Goldleaf")
+    latest_release=$(./common.sh get_latest_release "${2}" "XorTroll" "Goldleaf" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -136,7 +145,7 @@ download_goldleaf () {
 
 download_kosmos_toolbox () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "AtlasNX" "Kosmos-Toolbox")
+    latest_release=$(./common.sh get_latest_release "${2}" "AtlasNX" "Kosmos-Toolbox" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -150,7 +159,7 @@ download_kosmos_toolbox () {
 
 download_kosmos_updater () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${3}" "AtlasNX" "Kosmos-Updater")
+    latest_release=$(./common.sh get_latest_release "${3}" "AtlasNX" "Kosmos-Updater" "0")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -164,7 +173,7 @@ download_kosmos_updater () {
 
 download_ldn_mitm () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "spacemeowx2" "ldn_mitm")
+    latest_release=$(./common.sh get_latest_release "${2}" "spacemeowx2" "ldn_mitm" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "ldn_mitm*" "*.zip")
     file=$(./common.sh download_file "${asset}")
@@ -178,7 +187,7 @@ download_ldn_mitm () {
 
 download_lockpick () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "shchmue" "Lockpick")
+    latest_release=$(./common.sh get_latest_release "${2}" "shchmue" "Lockpick" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -191,7 +200,7 @@ download_lockpick () {
 
 download_lockpick_rcm () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "shchmue" "Lockpick_RCM")
+    latest_release=$(./common.sh get_latest_release "${2}" "shchmue" "Lockpick_RCM" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "*.bin")
     file=$(./common.sh download_file "${asset}")
@@ -204,7 +213,7 @@ download_lockpick_rcm () {
 
 download_sys_clk () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "retronx-team" "sys-clk")
+    latest_release=$(./common.sh get_latest_release "${2}" "retronx-team" "sys-clk" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "sys-clk*" "*.zip")
     file=$(./common.sh download_file "${asset}")
@@ -238,7 +247,7 @@ download_sys_ftpd () {
 
 download_nxdumptool () {
     mkdir -p ${1}
-    latest_release=$(./common.sh get_latest_release "${2}" "DarkMatterCore" "nxdumptool")
+    latest_release=$(./common.sh get_latest_release "${2}" "DarkMatterCore" "nxdumptool" "1")
 
     asset=$(./common.sh find_asset "${latest_release}" "nxdumptool*" "*.nro")
     file=$(./common.sh download_file "${asset}")
@@ -247,6 +256,26 @@ download_nxdumptool () {
     mv ${file} "${1}/switch/NXDumpTool/NXDumpTool.nro"
 
     echo $(./common.sh get_version_number "${latest_release}")
+}
+
+remove_configs () {
+    # Atmosphere
+    rm -f "${1}/atmosphere/BCT.ini"
+    rm -f "${1}/atmosphere/loader.ini"
+    rm -f "${1}/atmosphere/system_settings.ini"
+
+    # Hekate
+    rm -f "${1}/bootloader/patches.ini"
+    rm -f "${1}/bootloader/patches_template.ini"
+
+    # System Modules
+    rm -f "${1}/config/hid_mitm/config.ini"
+    rm -f "${1}/config/sys-clk/config.ini"
+    rm -f "${1}/config/sys-ftpd/config.ini"
+
+    # Apps
+    rm -f "${1}/switch/KosmosToolbox/config.json"
+    rm -f "${1}/switch/KosmosUpdater/settings.cfg"
 }
 
 # =============================================================================
